requirements.txt
protobuf
pydantic
asyncpg
python-dotenv
sqlalchemy
yt-dlp
fastapi
uvicorn[standard]
python-multipart
pymilvus
langchain
langchain-community
sentence-transformers
pypdf
youtube-transcript-api
datasets
transformers
ragas
pymilvus[milvus_lite]
google-generativeai
pytube
evaluate
langchain-openai
openai
yt_dlp
requests
pysrt
google-genai